controllerProduct

+++ IberaDelivery/Models/LnOrder.cs
@@ -5,11 +5,11 @@ namespace IberaDelivery.Models
 {
     public partial class LnOrder
     {
-        public long NumOrder { get; set; }
-        public long NumLine { get; set; }
+        public int NumOrder { get; set; }
+        public int NumLine { get; set; }
         public int Quantity { get; set; }
         public decimal TotalImport { get; set; }
-        public long RefProduct { get; set; }
+        public int RefProduct { get; set; }
 
         public virtual Order NumOrderNavigation { get; set; } = null!;
         public virtual Product RefProductNavigation { get; set; } = null!;
