+++ IberaDelivery/Models/Order.cs
@@ -10,10 +10,10 @@ namespace IberaDelivery.Models
             LnOrders = new HashSet<LnOrder>();
         }
 
-        public long Id { get; set; }
+        public int Id { get; set; }
         public DateTime Date { get; set; }
         public decimal Import { get; set; }
-        public long UserId { get; set; }
+        public int UserId { get; set; }
 
         public virtual User User { get; set; } = null!;
         public virtual ICollection<LnOrder> LnOrders { get; set; }
